src/basepairmodels/requirements.txt
tensorflow-gpu==1.14
tensorflow-probability==0.7
keras==2.2.4
scikit-learn
scipy
scikit-image
deepdish
h5py
numpy==1.16.4
pandas
matplotlib
plotly
deeptools
pyfaidx
modisco
deeplift
tqdm
pycodestyle
pytest
pandas

